+++ Natasha/ExternalAPIs/Kokoro/TimesPlayed.cs
@@ -1,0 +1,12 @@
+﻿namespace Shockah.Natasha;
+
+public partial interface IKokoroApi
+{
+	public partial interface IActionApi
+	{
+		CardAction MakeTimesPlayedVariableHintAction(int cardId);
+		IConditionalActionApi.IIntExpression MakeTimesPlayedCondition(int currentTimesPlayed);
+
+		int GetTimesPlayed(Card card);
+	}
+}